Persist TranscodingJobs across executions
This migrates to Java 7's Path and nio.File objects where possible.  The
AWS SDK still works with java.io.File from Java 6 and earlier.

Use DynamoDB to persist TranscodingJobs between executions.  Leverage
the records to enable input and output Path objects local to the system.
The exit condition of the Launcher will not work until a size() method
is added to the TranscodingJobRepository to detect that listening to the
SQS queue is no longer necessary.

Introduce TranscodeEvent as a new DTO to return information from the ETS
submission, since ETS messages do not contain the full local Path that
TranscodingRequest and TranscodingJob have.  Leverage the S3 Key-like
nature of the fields in TranscodeEvent to construct Movie objects to
manipulate the S3 bucket after transcoding is done.

Also adds a simple batch file to prevent having to remember how to
invoke Maven on the command line.

Next steps:
- fix the exit condition
- handle Ctrl-C gracefully
- test (and potentially code) wildcards on the command line